Extract TrelloAttachment into trello_attachment.dart
